# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Zentrivex Crypto Investment Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /withdrawals:
    get:
      operationId: listWithdrawals
      tags: [withdrawals]
      summary: List user withdrawals
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Withdrawal"
    post:
      operationId: createWithdrawal
      tags: [withdrawals]
      summary: Submit withdrawal request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
  /kyc:
    get:
      operationId: getKyc
      tags: [kyc]
      summary: Get current user KYC status
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Kyc"
    post:
      operationId: submitKyc
      tags: [kyc]
      summary: Submit KYC documents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KycInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Kyc"
...
  /admin/withdrawals/{id}/approve:
    patch:
      operationId: approveWithdrawal
      tags: [admin]
      summary: Approve withdrawal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
  /admin/withdrawals/{id}/reject:
    patch:
      operationId: rejectWithdrawal
      tags: [admin]
      summary: Reject withdrawal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
...
  /admin/kyc/{id}/approve:
    patch:
      operationId: approveKyc
      tags: [admin]
      summary: Approve KYC
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Kyc"
  /admin/kyc/{id}/reject:
    patch:
      operationId: rejectKyc
      tags: [admin]
      summary: Reject KYC
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Kyc"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, email, firstName, lastName, role, balance, kycStatus, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, admin]
        balance:
          type: number
        kycStatus:
          type: string
          enum: [none, pending, approved, rejected]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    Withdrawal:
      type: object
      required: [id, userId, amount, walletAddress, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        amount:
          type: number
        walletAddress:
          type: string
        network:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        rejectionReason:
          type: ["string", "null"]
        user:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
          format: date-time
        approvedAt:
          type: ["string", "null"]
    WithdrawalInput:
      type: object
      required: [amount, walletAddress]
      properties:
        amount:
          type: number
        walletAddress:
          type: string
        network:
          type: string
    Kyc:
      type: object
      required: [id, userId, documentType, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        documentType:
          type: string
          enum: [passport, drivers_license, national_id]
        frontImage:
          type: ["string", "null"]
        backImage:
          type: ["string", "null"]
        selfieImage:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        rejectionReason:
          type: ["string", "null"]
        user:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
          format: date-time
        reviewedAt:
          type: ["string", "null"]
    KycInput:
      type: object
      required: [documentType]
      properties:
        documentType:
          type: string
          enum: [passport, drivers_license, national_id]
        frontImage:
          type: string
        backImage:
          type: string
        selfieImage:
          type: string
    RejectInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
...