# lib/api-zod/src/generated/types/kycInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Zentrivex Crypto Investment Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { KycInputDocumentType } from './kycInputDocumentType';

export interface KycInput {
  documentType: KycInputDocumentType;
  frontImage?: string;
  backImage?: string;
  selfieImage?: string;
}